requirements.txt
gunicorn
django-heroku


cmsplugin-filer>=1.1





aldryn-bootstrap3==1.2.2
aldryn-style==1.1.4

dj-database-url==0.4.2
Django==1.8.18
django-appconf==1.0.2
django-classy-tags==0.8.0
django-cms==3.4.5
django-filer==1.3.0
django-formtools==2.1
django-mptt==0.8.7
django-polymorphic==1.3
django-sekizai==0.10.0
django-treebeard==4.2.0
djangocms-admin-style==1.2.7
djangocms-attributes-field==0.3.0
djangocms-column==1.7.0
djangocms-googlemap==1.1.1
djangocms-installer==0.9.7
djangocms-link==2.1.2
djangocms-snippet==1.9.2
djangocms-style==2.0.2
djangocms-text-ckeditor==3.5.1
djangocms-video==2.0.4
easy-thumbnails==2.4.2
html5lib==0.9999999
olefile==0.44
Pillow==4.3.0
pytz==2017.3
six==1.11.0
tzlocal==1.5.1
Unidecode==0.4.21

